Use _exit() on fatalError, and better clean up where we still use exit().
ASAN's leak detection may not terminate if the fiber stack is still
around at termination time. We ran into this when fatalError() would
call exit() when rt::done() hadn't be executed yet; and also with a few
manual exit() calls we still had that weren't shutting down the runtime.

So we now do a hard terminate with _exit(); we reduce the uses of
fatalError() further to just a tiny set of really catastrphic runtime
libraary issues; and we clean up before exits().