Add runtime checks for stack size.
When the HILTI codegen creates a function, it now always inserts a stack
check at the beginning that ensures that a minimal amount of stack space
remains available. If it is not, we abort with a runtime exception. This
gives us well-defined behavior if call stacks go to deep.

The minimum amount of stack space is configurable through a global
option. Picking the right value is a bit of black magic: it seems that
it needs to be sizeable for the program to actually recover, presumably
because throwing the exception needs quite a bit of space itself. But we
can be generous here: thanks to having a shared fiber stack now,
budgeting a bit more doesn't cost us much.

Addresses #460.