Allow to set a compiler wrapper through environment.
By default, our JIT uses the same C++ compiler as HILTI/Spicy were built
with themselves. The environment variable HILTI_CXX can override that
to switch to a different executable to use. It won't work to switch
in a completely different compiler that way (because we'd continue to
make some assumptions based on the default compiler), but it allows to
provide an additional wrapper around the default compiler.