Extend {hilti,spicy}-config with options to print compiler flags for building HLTO files.
This adds --cxx-flags-hlto and --ld-flags-hlto to both tools. It's
mostly for completeness and also debugging, we aren't using it anywhere
right now.

Also includes some cleanup on where options are defined. In particular,
we move some standard JIT options out of CMake variables into
``config.cc.in` for easier comprehension.