Make test output deterministic.
The order in which modules are registered is non-deterministic since for
compiled libraries we rely on the dynamic library loder to perform the
registration. This in turn makes the order in which any code at module
scope is executed at load time non-deterministic as well.

Fix the test `hiltic.cc.separate-units` by sorting its output.