Use "default" compiler if possible in Docker images.
Since we now compile directly with the system compiler we do not
strictly require LLVM anymore. This patch simplifies the in-tree Docker
images so they use the default compiler for the platform if possible;
since we require a C++17-capable compiler we still need to install
custom compilers on a number of platforms.