Allow to enforce sequential execution of jobs spawned during JIT.
By default JIT spawns compilation jobs in parallel to optimize
throughput. This can lead to issues when working with hard resource
constraints, e.g., in containers, as peak memory is driven by the hidden
number of launched compilation jobs is hard to control without knowing
exactly which jobs will be launched.

This patch modifies the JIT code so that if the environment variable
`HILTI_JIT_SEQUENTIAL` is set jobs are executed sequentially.