Allow to pass mime type to data_in when sending data to Zeek.
This change adds zeek::file_data_in_with_mime_type - which allows to
pass an initial mime type to Zeek when sending file data.

I also added the timestamp removal to the Zeek log canonification. This
is needed for the testcase - and it also seems probable that this will
be needed for future tests. It also is used for nearly all btests in the
Zeek source tree.