Publish a single Docker image.
This patch moves us from publishing a number of Docker images for
different platforms to a single image. This simplifies the choices users
have who want to spin up a Spicy container, but do not necessarily care
about the underlying distribution.

We still maintain Dockerfiles for a choice of platforms in tree.