Fix parsing of single-element container constructors.
Some parts of the parsing of container constructors are reusing parsing
of tuples which lead to issues when constructing single-element
containers, e.g.,

   let x = vector<uint16>(16);

failed to parse since no constructor was parsed while e.g.,

   let x = vector<uint16>(16, );

succeeded.

This patch fixes the parsing of tuples so that the trailing `,` is
optional as intended.