Add support for shared unit contexts with lifetime determined by host application.
This introduces a new unit property for public unit types:

    %context = <type>;

The unit's new `context()` method will then return a value of type
`<type>&` (i.e, a reference to an instance of the given type). This will
be initialized with a reference to a default-constructed instance of the
type.

What makes this special is that the host application can determine how a
context instance is being shared across different parsers: it can (1)
let multiples units link their context to the same instance, and (2)
retain a context instance beyond any specific unit's lifetime.

We use this new mechanism to provide contexts with connection-level
semantics: the Zeek plugin, and also spicy-driver's batch mode, will
create a single context instance that the two originator-side and
responder-side units will share (they must declare matching contexts
for that to work). The object will live until the connections gets
torn down (also meaning it'll be reused across the per-packet parsers
for UDP).

An example how to use this for associating replies with requests in
a pipelining protocol:

    type Connection = tuple<pending: vector<bytes>>;

    public type Requests = unit {
        %context = Connection;
        : (Request(self))[];
    };

    public type Replies = unit {
        %context = Connection;
        : (Reply(self))[];
    };

    public type Request = unit(inout reqs: Requests) {
        cmd: /[A-Za-z]+/;
        : b"\n";

        on %done {
            reqs.context().pending.push_back(self.cmd);
            }
    };

    public type Reply = unit(inout replies: Replies) {
        response: /[0-9]+/;
        : b"\n";

        on %done {
            if ( |replies.context().pending| ) {
                print "%s -> %s" % (replies.context().pending.back(), self.response);
                replies.context().pending.pop_back();
                }
            else
                print "<missing request> -> %s", self.response;
        }
    };

Feeding spicy-driver with a litte batch file containing a couple
connections, each having a series of request/reply pairs:

    msg -> 100
    put -> 200
    CAT -> 555
    end -> 300
    get -> 400
    LST -> 666

Closes #696.