Add new variant of external function linkage to HILTI that skips the resumable wrapping.
Linkage `external-no-suspend` can now be used to expose functions to
external C++ code that are guaranteed to never suspend. Doing so skips
the overhead of wrapping the call into a fiber resumable, and also
allows to carry through the original signature of the function.

Includes a fix for calling an `extern` function from inside the same
module. The new test case triggered this.