Support finalizers in HILTI structs.
There's a now new magic hook `~finally` that will run automatically
when a struct instance gets destroyed:

    type X = struct() {
        hook void ~finally();
    };

    method hook void X::~finally() {
        hilti::print("finally running!");
    }

The exact timing for when the hook runs remains undefined. On the C++
side, this runs from the struct's destructor, but due to reference
counting it can be hard to predict when exactly that will happen.

A finalizer may also run for more instances that one would intuitively
expect because a struct may get default constructed or copied in
non-obvious places.

I'm not very stronly tied to the name `!finally`, just couldn't think
of much else under the assumption that (1) the name should be static and
independent of the struct type, (2) not be a legal identifier otherwise,
and (3) not match Spicy's '%'-named hooks to avoid naming conflicts.

Note: This was originally meant to support the Spicy-side finalizer.
Turns out that doesn't work well because it's too easy to trigger cycles
where temporary structs of the same type trigger finalizer again. I'm
still leaving this is because I can see other uses for this
functionality in the future.