Fix Exif parsing
The current state is still WIP - but, in principle, Exif parsing now
works for both, png and jfif. This also exposed a few bugs on the way.

However, there are a few snags that I want to fix - the way that the
fact that there can be a next_ifd is handled at the moment is
suboptimal.

Commit because this state can be used to trigger GH-740