Prepare unit field AST node for upcoming changes.
This mainly lifts the requirement that a vector element cannot be
another field/item. Doing so requires making the typing more dynamic.

This is in preparation for upcoming changes to the parser generator,
won't compile on its own.

(Note: the change adding hooks_start/hooks_end members to the node
turned out to be unnecessary, but leaving that in place as it seems the
right thing to do and might end up being useful in the future if we
added more child nodes.)