Support parsing of nested fields.
This is main commit in this series: it lets the parser builder support
arbitrary nested fields by recursively going through their structure.
The change comes with some major refactoring & cleanup to make that work
(and might hence be difficult to review.)

The new test case shows a number of examples that are now supported:

    public type Test = unit {
         i1: (bytes &size=5);
         i1b: (bytes &size=5 &convert=42);
         i2: (int8[3]);
         i3: (int8[2])[3];
         i4: (int8[] &until=($$=='X'))[] &until=($$==vector(55, 56));
         i5: (FooImpl);
         i6: (FooImpl[2]);
         i7: (FooImpl[2])[3];
         i8: (FooImpl &convert=42);
         i9: (FooImpl &convert=42) &convert=$$*2;
       i10a: (bytes &size=5 &convert=$$.to_int());
       i10b: (Bytes &size=5 &convert=$$.to_int());
       i10c: (FooImpl &convert=$$.data.to_int());
        i11: (FooImpl &convert=$$.data.to_int()) &convert=$$*2;
        i12: (FooImpl &convert=$$.data.to_int())[2];
        i13: (FooImpl &convert=$$.data.to_int())[2] &convert=[3 * xyz for xyz in $$];
    }

While the expectation isn't really that people would write this, this
puts the infrastructure in place to support functionality that needs
more complex field structures. The first application will come in the
subsequent commit.