Add new function linkage "preinit" to HILTI.
A "preinit" function will execute immediately when the code is loaded,
even before the runtime environment gets initialized. In contrast,
normla "init" functions only register themselves at that point, and then
execute later when the runtime starts up.

This allows a "preinit" function to trigger functionality early on, but
it will not have the runtime environment available yet, so pretty much
any use of HILTI funtionality is ill-defined at this point. The only
thing such a function can do it manipulate state outside of HILTI.

In the Zeek plugin, we will use this to call into Zeek early on.