Fix bug with custom enum types not being available in time for Zeek scripts.
Turns out that for precompiled Zeek analyzers it didn't work to
automatically make the Spicy-side enum types available to Zeek scripts
because of a timing issue: the plugin would register them only at
analyzer initialization time, which is after other scripts were already
parsed, leading to errors about unknown IDs if they were trying to use
the enums. (This worked for JIT analyzers compiled by Zeek itself
because of an additional early registration in that case, which is why
we hadn't caught it yet).

Fix is to use the new "preinit" function to do the enum registration
early.

We now also add the enum type to the output of `zeek -NN`.