Let unresolved unit fields always track the original ID.
So far we had the ID resolver potentially replace the original ID. Turns
out that introduced an ordering issue: depending on what it got
replaced, the Spicy-side ID resolver could see different things, and
hence proceed differently. Now we give the ID its own children slot.
(The for which this matters will come in a subsequent commit.)