Fix parse error message when field requirement fails.
The previously generated error message assumed that field `&requires`
clauses contained just a value assertion, e.g.,

    x: uint8 &requires=($$ == 2);

generated

    &requires failed ($$ == 2)

If `&requires` referenced a function $$ -> bool this would have always
printed the return value of the function which is not useful.

This change brings the parse error message for fields in line with the
one we already generate for unit `&requires` clauses.