Factor preprocessor logic out of EVT parser.
This makes the @if/@else/@endif logic reusable through a utility class
coming with the HILTI toolchain library.

Couple of extensions included as well:

    - Undefined IDs are now always treated as if they were set to zero.

    - We now support negation (`@if ! ...`) and `@if <id>` as a short
      cut for `@if <id> != 0`.

It also fixes processing of nested if/else-chains, which was broken
(turns out an existing test had recorded a wrong deadline here as well).

Tests to come in a subsequent commit.