Fix installation of Zeek plugin.
In order to install the Spicy Zeek plugin we now delegate all work to
the existing Zeek plugin packaging infrastructure. It already deals
correctly with e.g., `DESTDIR`-style installations so is a better
candidate to implement this than a homegrown solution.

We will change the CPack setup to enable `DESTDIR` in a subsequent
patch.