Update instructions on binary artifacts.
Since some Spicy binaries have baked in paths we always required
installation to the configured prefix, and tarballs of binary artifacts
needed to be extracted in exactly matching locations.

This patch updates the installation instructions to reflect that. Not
being able to install into a dedicated prefix means that users need to
in some record what files where installed. We give a small recipe for
that.

Ideally we would be able to set a `CPACK_PACKAGING_INSTALL_PREFIX` like
`/opt/spicy` and have that reflected in the binaries and installation
paths. Sadly our current installation setup does not support such
setting right now.