Fix release workflow to also upload deb and rpm files.
While the action we used previously did support globs for input files,
it still required specifying an output file name which we cannot give
since it might be one of tar.gz/deb/rpm. Additionally, the action is now
unmaintained. Due to this we change to a different action in this patch.