Use more recent CMake in some containers.
In order for CPack with a DEB target to correctly deal with the symlinks
we use for some 3rdparty modules we need CMake's
fb5d24fa3ca5061bbae9e68a6e27636e7366a271. We only plan to create DEB
files on Debian-based systems, but update the versions where possible.

Building CMake on Alpine Linux seems to be some work, so we punt on
actually increasing `cmake_minimum_required`.