Fix vector parsing potentially missing EOD.
When parsing a vector bounded by &eod, the parser could enter the
code for reading the next element too early before it could reliably
tell that EOD hadn't been reached yet. The fix is to change `atEod()` so
that if stream is still unfrozen, it waits for at least one available
byte before making the decision.

This is hard to test for because it requires yielding & resuming the
input in the right way, but the PNG can act as a test case: it was
producing an unexpected weird, which is now gone.

Closes #817.