Drop noexcept in Sink move constructor.
Since `Sink` contains a `std::list` which is not required to have a
`noexcept` move constructor, it might be impossible for the default'ed
move constructor for `Sink` to be `noexcept`.

While many platforms seem to provide a `noexcept` `std::list` move
constructor, this is e.g., not the case on macOS Mojave.