Fix regexp matching across chunk boundaries.
We previously were passing incorrect information to JRX on whether a
block in a stream was the final block. This lead to incorrect matches
when the pattern to match was split across a block boundary and a match
was already possible with the first block. This manifests e.g., in the
following Spicy code when passed the input `x\n y\n`:

```
module Test;

public type X = unit {
    on %init { self.connect_filter(new F); }
    x: b"x";
    : /[ \n]*;
    y: b"y";
};

type F = unit {
    %filter;
    x: /[^\n]*\n/ { print "x1", $$; self.forward($$); }
    y: /[^\n]*\n/ { print "y1", $$; self.forward($$); }
};
```

Here the filter leads to the input to be split into two blocks `x\n` and `
y\n`. When attempting to parse the dummy field `/[ \n]*` we consumed only the
`\n` from the first block, but not the leading space in the second block.

We now pass correct information on whether a block is the last one the JRX.

Closes #860.