Fix version string checking out release tags.
The script to autogenerate version strings would previously only detect
releases when run from a release branch. Since our CI usually checks out
release tags this would have lead to incorrect versions strings.

We now detect if we are computing the version string for a tag and
handle it separately.