Do not automatically create `spicy:latest` Docker tag from `main`.
We currently do not automate detecting whether a certain tag corresponds
to the latest release (i.e., whether it is equivalent to the `release`
tag). In any case, With the release of spicy-1.0.0 we should create the
`spicy:latest` image from a dev version anymore, so we remove that
automatism with this patch.