Fix handling of `&size` attribute for anonymous void fields.
We previously did not update the parser state in productions for `void`
variables. This e.g., lead to `&size` checks to always fail when used
with `void` variables, even though we correctly advanced in the input
stream.

With this patch on can now specify blocks of data of arbitrary size to
skip over, e.g.,

    type X = unit {
        : void &size=2; # Skips two bytes.
        a: unit8;
    };

Closes #37.