Lower minimum required Python version to 3.2.
We raised this to python-3.6 be able to use newer features in
scripts/autogen-version, but it turns out that we need to support down
to at least python-3.5 for platforms like ubuntu-16 or debian-9.

We now go all the way to python-3.2 which is very conservative. For that
we got rid of uses of subprocess.run (since 3.5) and f-strings (since
3.6).