Detect presence of sink hooks at compile time.
Sink hooks are by default generated for any `public` unit. With the
introduction of an optimizer pass to remove unused hooks we might now
remove them and need to detect their presence when e.g., registering a
parser.

This patch moves from these instances from branching on whether a unit
is `public` to directly detecting the presence of the hooks.