Make sure container sizes are runtime integers.
For the HILTI runtime containers we previously directly forwarded the
`size` functions of the underlying stdlib containers. This could have
lead to them returning a type not expected by our runtime, e.g., on
macos these might return `unsigned long` values which our runtime e.g.,
doesn't know how to print.

This patch changes these functions so that now they return types which
we expect our runtimes to be able to work with.