Do not update unchanged version files.
When porting this file from shell to Python we lost the feature that
outputs would not be emitted if they were unchanged. After the port we
always emitted the files which leads to potentially expensive but
unneeded regeneration of consumers of this file.

This reintroduces a check which skips emitting output files if they
would have been identical to existing files. This prevents updating of
timestamps which allows CMake to skip regeneration of consumers.