Default-initialize forwarding fields without type arguments.
The code to default-initialize fields previously default-constructed the
field with any arguments given. With the introduction of unit `&convert`
expressions this might produce incorrect code, e.g., for cases where the
unit can be constructed with parameters but has a `&convert` expression
which changes the type. Additionally, the code here constructed a
default value of the final type (after `&convert`) so the invoked
constructor might not even be applicable.

Since we generate units with a default constructor forwarding field
parameters to any constructor is not needed, so with this patch default
constructor values to default-initialize fields with.

Closes #917.