Remove accidental truncating conversion.
The intention of this code was to compute a mask for an `uint64_t`, but
accidentally performed to compuation with `unsigned long`. This
potentially lead to shifting beyond the width of the input value (e.g.,
for `range==63`).

We now enforce using wide enough types to remove this potential UB.