Adjust processes used to build Docker image.
Github action runners only have limited resources (currently for Linux:
2 CPUs, 7GB RAM). In the build of the spicy-plugin and spicy-analyzers
Zeek packages we previously hardcoded a certain level of parallelism,
which caused us to run into especially RAM limits when building
spicy-analyzers. With recent versions of these packages we can now
optionally externally parameterize the level of parallelism to use.

This patch adjust the Github action setup so we build these packages
with a number of processes compatible with the runners we will be faced
with.