Use forward declarations in linker metadata.
When computing the `aux_types` of hook parameters for linker joins we
previously used whatever cxx declaration the `CodeGen` had stored which
could have been a forward declaration or a full declaration. Since
`CodeGen::typeDependencies` does currently not reliably compute the list
of type dependencies, see #949, we could encounter incomplete lists of
types in linker joins if a full, not a forward declaration was emitted.

With this patch we explicitly compute a forward declaration which is
sufficient since user-land Spicy types are passed as `ValueReferences`.

Closes #918.