Avoid name collision when codegen'ing struct element initializers.
When codegen'ing a struct constructor which can initialize all fields
from passed in values we previously distinguished names of our arguments
from the user-specified fields by appending a `_` to our arguments. This
could have lead to name collisions if user code contain field names
differing only in whether they contained a final `_`.

With this patch we take a different approach in that we now explicitly
qualify access of fields with `this->`.