Add runtime check for parsing progress during loops.
When parsing a vector doesn't consume any input during a loop iteration,
we now abort with a parse error to avoid infinite loops (because next
time we'd very likely only do the same nothing once more, and then
again, and again, ...).

Strictly speaking, there could be a legitimate reason for sometimes not
making progress, but that seems obscure enough to enforce this
restriction in favor of robustness. Our existing analyzers don't have a
problem with it. If it turned out it's really needed for something, we
could add an attribute to disable this check.

Closes #940.