Remove spicy-plugin and spicy-analyzers submodules.
We no longer support building them in-tree. This comes along with
removing most Zeek-related pieces from this repository.

Because we can't auto-generate Zeek-related references docs, there's now
a new scripts `scripts/autogen-zeek-docs` that needs to be called
manually to do so. (Eventually, we may move the Zeek part of the docs
over to the plugin repository.)

This also needs to remove the `spicy.lib.protocols.{dns,http}` tests
because we don't have the source files anymore.