Add unique "linker scope" identifier to HLTO files.
This can be used to disambiguate where code comes from.

In Spicy, we use now store this linker scope along with the parsers
when they get registered with the runtime. In a parallel commit, the
Zeek plugin then uses that to limit the parsers it uses to those from
the same HLTO unit as the EVT file currently being processed.