Make detection of current executable more robust.
We previously assumed that we could always detect the path to the
current executable. It turns out, that the way the PathFind library we
are using is written that it can fail to return a result especially on
FreeBSD; in failure case an empty path is returned.

This patch adds failure handling for PathFind so the empty path is not
propagated up to the caller. This gives us an option to emit a dedicated
and probably more helpful error message. We also add special handling
for FreeBSD with instructions for users on how to recover.

Closes #961.