Force recomputation of version information if generating VERSION file.
We previously would always used persisted version information, even when
generating a fresh `VERSION` file. This meant that while e.g., our CMake
setup generated a fresh `VERSION` file on each configure run, that
information never got updated. This could have lead to inconsistent
version information, especially in iterative development.

We now largely ignore an existing `VERSION` file if the `--store` flag
is passed to `scripts/autogen-version` so that in that case the version
information can be updated. In order for the version script to work in
release tarballs (which should not have git version information) we allow
a soft fallback to information from the `VERSION` file, even if asked to
recompute, but the recomputation failed.