Explicitly mark parse functions of filters as `&always-emit`.
Filters expose as part of their API their parse functions to e.g.,
`sink::connect_filter` without the filters needing to be `public`.

In order to remove unneeded functions in the global optimizer in a
follow-up commit we mark the parse functions of filters as
`&always-emit` in this patch. This will allow us to use that information
in the optimizer to prevent removal of these functions, even though
their uses are not visible in the Spicy AST.