Swallow Git error messages during CMake if run from tarball.
If we run our current CMake configure setup from a release tarball, we
cannot detect whether the release tarball corresponds to a Git tag.
While this should be fine for users, it still caused us to leak a Git
error message from the CMake run.

With this patch we now swallow the error message to be less confusing
for users.

Closes #960.