Ensure cleanup of temporary driver HLTO files on errors.
If we are executing code we just compiled we need to clean up the
generated temporary HLTO file. The file is held in a RAII container
so any needed cleanup happens automatically on destruction.

This patch adds explicit triggering of any cleanup so we can continue to
directly call `::exit` on error paths.