Annotate generated parser code with feature flags.
This patch adds module-level boolean constants which signify which
features to enable for which units which currently are all set to true.
When building a parser we can then guard features-dependent code behind
a check for the feature flag.

Currently this only leads to slightly more code being emitted. In
follow-up commits we will add an optimizer pass which can detect unused
features and disable them. The feature-dependent code can then be
removed by constant folding.