Prepare `statement::Block` for optimizer edits.
In a subsequent commit we will introduce optimizer changes which can
perform edits on `statement::Block`. Like in all other optimizer passes
we will model removals of nodes as replacements with none nodes.

This patch prepares `statement::Block` for these changes so it does
assert anymore that all its children are statements.