Change return type of `AttributeSet::add` so it always returns some set.
Thes functions add an element to a given set (which could be none), and
return a set with the new element present. That means that the new set
is always some value.

This brings the signature of the overload taking an `optional` set in
line with the one taking a naked set; both functions now return a set.