Change optimizer tests to compare HILTI code instead of ASTs.
We previously checked diffs of ASTs. This was both noisy and at times
hard to understand. Since we now generate HILTI output after
optimizations we can change the tests to instead output HILTI code. We
in general output two source files, one with and one without
optimizations. To review changes it is still required to diff these
files somehow, but we do not directly check in diffs anymore as
reconstructing the original, unoptimized code from the diff wasn't
always very clear.