Add optimizer pass removing unused `%random-access` features.
We previously added global constants signifying which features a
generated parser supported, and added annotations to generated struct
fields related to `%random-access`.

All methods related to `%random-access` are modelled as builtin
operators for which we directly inline their definitions into the AST
during transformation. This means that method calls like `offset()` or
`position()` are not present in the AST. In order to detect user-code
uses of these methods we check accesses to `%random-access`-related
fields outside of feature-dependent conditional blocks (we expect
accesses inside feature-dependent blocks to e.g., update the fields).