Hardcode DockerHub org we publish images for in GH workflow.
The DockerHub org we publish to previously matched the user we used to
log in to DockerHub, so we could avoid hardcoding the full org name.
Since we now use a bot account to publish to DockerHub this is not
true anymore and we need to specify the org explicitly.

This patch hardcodes the org name as `zeekurity` into Github workflows
publishing images to DockerHub.