Fix generated code to always invoke parsing of subunits.
When we introduced feature guards in the parser builder we wrapped
feature-dependent code in the guard. In this particular instance this
could have lead to semantically incorrect code since here `filter_init`
was used to detect the presence of filters, but the code was wrapped in
a generated conditional; this might have lead to situations which
disabled both the `true` and `false` of the conditional so subunits were
never properly parsed.

We now move the non-filter branch out of the feature conditional so at
it cannot be optimized out. This also removed some duplication in the
parser generator code.