Always emit `__sink` field for units with `%mime-type`.
If a unit has a `%mime-type` property it can be connected to via
`Sink::connect_mime_type` by MIME type string. In the optimizer we have
no way to connect such uses back to the type.

With this patch we now set the `__sink` field of such units
`&always-emit` so it never becomes a candidate for removal.