Properly track use of uninitialized enum values in units.
When analyzing whether a type was used we previously missed uses of
uninitialized enum values in unit variables, e.g., the following use was
missed

    type E = enum { a, b, c };

    public type U = unit {
        var e: E; # Uninitialized; missed.
    };

This could have lead us to incorrectly remove the enum type as unused.

With this patch we now properly track such uses as well. We also extend the
test suite to capture more uses of enum types.

Closes #979.