Disable constant folding for anything but feature variables.
We previously would inline init expressions for global variables, even
though the values of globals could have later been updated. Ideally
these feature flags should be constants so we know they cannot be
updated on the script side, but this is currently not straightforward,
see #982.

This patch disables constant folding for anything but our feature flags
which we control and from which we know that they cannot be updated at
runtime (their IDs cannot be used in userland Spicy scripts). As a
workaround for testing we also treat names starting with `feat________`
as feature flags; uses of globals with such names can still show the
problematic runtime behavior. We should rewrite our feature flags as
`Constants` in a follow-up patch.