Disable leak tests if `ASAN_OPTIONS` are already set.
These tests need to override the `ASAN_OPTIONS` environment variable to
work. If the setup already requires them to be set to something else it
isn't really clear what we can set in them here and we should skip the
test; e.g., on osx we require `ASAN_OPTIONS=detect_leaks=0` since leak
detection is not supported there.