Precompile `libspicy.h`.
Since only a single precompiled header can be explicitly added we need
to add this header to Spicy's `CXXFLAGS` differently in that it replaces
any precompiled header for `libhilti.h` we inherited from the HILTI
config.

We also include `libhilti.h` in `libspicy.h` now so we still get the
precompilation benefits for that part as well.