Unify `transform` and `filter` functions.
This patch unifies the various `transform` and `filter` functions in
hilti and hilti-rt.

`filter` is now implemented once in pretty straightforward generic code.
`transform` on the other hand is slightly more work as it requires
computing a target container type. We capture that in a function, and as
a cleanup get rid of overloads changing the container type (e.g., we
previously transformed a `list` into `vector`); since we cannot
generically compute a non-default allocator for a new vector we keep the
type change from a vector with custom allocator to a vector with default
allocator though.

Closes #999.