More fixes for building on ARM64.
In 89d3295f1363d8d07eda5c711740b18f31483d88 we adjusted some
preprocessor code to enable building on arm64. That patch missed an
update of another instance to the same conditional which we fix with
this patch.

Some ARM64 platforms do not seem to reliably set the `__arm64__` define,
but instead `__aarch64__`. We now allow that as an alternative.