Use `unordered_set` instead of `set` for `Location` cache.
Creation of `Location`s is in the hot path of the parser so that
performing member checks on `set`s becomes a bottleneck.

With this patch we now use an `unordered_set` for the cache of
`Location`s which can perform better since it has weaker guarantees than
a `set`. For a trivial module (just containing e.g., `module foo;`) this
nearly halves the time to parse (we also parse the implicitly included
modules `hilti.hlt` and `spicy_rt.hlt`), at least in debug builds; it
should have a similarly beneficial effect on larger modules as well.