Use `build` instead of `emplace` Bison variant API.
The `emplace` API for Bison variants was only added in bison-3.2 which
is not available on e.g., centos platforms. This patch replaces uses of
it with `build` instead.

Since `build` deduces the target value in the variant from its argument
type we need to make sure all values passed to `build` have the correct
type. This is error-prone, so we should consider reverting this patch as
soon as we drop support for Bison versions before 3.2.