Fix overly generous generation of type information.
We recently started including type information for all types that a
module included. That doesn't harm functionally, but increased code
size and compilation time. This patch reverts that to the old behavior.

Doing so required one additional fix for exceptions, which had
generation of their destructor tied to the generation of type
information; which now doesn't work anymore, but also seems unnecessary
in the first place.

Addresses #1009.