Explicitly enforce that a module is only registered once.
We previously allowed registering the same `HiltiModule` multiple times.
Since modules are backed by shared libraries this lead to unclear
semantics on whether e.g., their initialization code was called once or
multiple times.

With this patch we now enforce that a module with a given name and ID can be
loaded exactly once. We here introduced a dedicated module ID derived
from the linker scope to not require module names to be unique.

WIP module id.