Model types in `set` and `map` `in` operators.
The signatures of these functions previously contains at least two
wildcards, one for the type of element to search and another one for the
matching value in the container type. We however never validated that
the types matched so that no argument type validation was performed and
even type coercions were only performed implicitly by the C++ compiler.

With this patch we now check the types of the involved operands by
potentially inserting an explicit coercion of the needle to the element
(key) type if needed.

Closes #994.