Fix build for RTD.
This resolves the build problems due to sphinx-doc/sphinx#9727 we've
been seeing. RTD seems to by default run a <sphinx-2.0 which conflicts
with newer docutils.

Since we usually install sphinx from the system for building the docs
instead pin docutils in the docs requirements file which is parsed by
RTD.