Fix potentially use-after-free in stringification of `stream::View`.
To construct `UnsafeConstIterator`s we previously directly used a
`SafeConstIterator`'s chunk pointer which might be invalid and point to
e.g., random memory. With this patch we use the safe API to retrieve it
which either yields a valid pointer to a chunk or a `nullptr` (which we
also handle).