Make `offset` return correct value even before parsing of field.
We previously would update `__position` only after beginning a
new production. This meant that if a user used its value before that,
e.g., to compute a field's `&size`, the old value was still visible.

With this patch we now update its value as soon as the previous
production has finished parsing and run its hooks.

Closes #1087.