Avoid trimming input too low in call stack.
We we extract an integer literal during regular parsing we cannot know
at the point where we unpack the bytes whether the extracted value
matches the expected value. This means that we should not trim the input
at that point either.

This patch moves the trimming to a point where we can know whether the
extracted value matches our expectations.