Fix simultaneous matching of regexps and non-regexps.
We previously generated two different synchronization loops from
`getLookAhead`, one directly there for regexps and one deeper down for
literals. This meant that we were not able to match all of these at the
same time (technically we would always try to match only the regexps).

This patch reworks the the code so we generate exactly one
synchronization loop in `getLookAhead`.